Obscure VL and std/strings updates
Added a simplified, logic-mode, guard-verified version of ACL2's
evisceration function, for use with str::pretty.

Updated the VL printer's ~x directive to eviscerate terms (with
reasonably large defaults).  This avoids blowups when printing certain
warnings that contain huge (printed) svex expressions.

Extended the VL string lexer so that it handles more escape sequences;
generally it seems that Verilog tools treat \A as A, \B as B, and so
on, for most characters that aren't special things like \n for
newline, etc.  Added a new "str" cosim to test this out somewhat.

Extended VL's preprocessor to handle a really weird feature that other
tools seem to implement.  Normally the special sequence `" is found
in `define bodies as a way to insert " characters into the expansion
of a macro.  It seems that other tools also weirdly allow it to be used in
the arguments to a define, e.g., `foo(`"bar`") seems to be allowed and
treated like a string.  Extended VL's preprocessor to be able to
handle this and added a few unit tests based on the observed behavior
of the simulators.