VL: fix obscure parsing bug with event expressions
The event_expression parser was not correctly looking for things like 'or' when
the first expression had extra parens around it, e.g., for something silly
like:

    always @(((posedge foo)) or ((posedge bar))) ...

We would stop matching after the end of 'foo))' and fail to parse the rest of
the event expression.  Fix that, and extend the tests to validate the fix.