Rename and add theorems for some term utilities.
The (PSEUDO-)FN/LAMBDA-P recognizers are now called (PSEUDO-)TERMFNP, where
'termfn' stands for 'term function', i.e. the functions that appear in
translated terms. (The FUNCTIONP symbol could not be used because it is among
the ones imported from Common Lisp.)

The topic TERM-APPLICAND-RECOGNIZERS has been renamed to
TERM-FUNCTION-RECOGNIZERS.

The new theorems show that the (PSEUDO-)TERMFNP and (PSEUDO-)LAMBDAP recognizers
correspond to the portions of (PSEUDO-)TERMP that constrains the functions in
the terms.

Thanks to Eric Smith for suggesting to use 'function' instead of 'applicand',
and to Matt Kaufmann for useful discussion.