Extend and improve world query utilities.
Now many world query utilities come in two variants: a "fast" one and a
"logic-friendly" one. The fast utilities are similar to the built-in system
utilities (which they complement), in that they have relatively weak guards and
do not have (strong) return type theorems. The logic-friendly utilities have
stronger guards and include believed-to-be-always-satisfied run-time checks that
allow us to verify their guards and to prove stronger return type theorems.

There are also a few logic-friendly variants of built-in system utilities.

Some of the documentation of these utilities has been revised for greater
consistency within this collection of utilities.

Some utilities have been reordered in the file, and separation lines of
semicolons have been added to group related utilities.