Resize-list is now tail-recursive.  Stack overflows are handled better now in LispWorks builds.  Modified random-list in books/centaur/misc/seed-random.lisp to use tail recursion.
Quoting :doc note-7-5:

  The function [resize-list] is now defined using [mbe] so that its
  execution is tail-recursive.  Thanks to Martin Simmons of LispWorks
  Technical Support for diagnosing a stall in the certification of
  community book books/centaur/truth/perm4.lisp as being due to a
  stack overflow caused by an invocation of resize-list.  Based on
  his advice we no longer automatically grow the stack in LispWorks;
  this will ease debugging when compilation is done with safety 3.
  The maximum stack size is 399998, at least in our 64-bit LispWorks
  build.

Also made a few small related edits.

NOTE: Now that LispWorks no longer increases the stack indefinitely
(until it dies or stalls), there was initially a stack overflow in
books/centaur/misc/seed-random.lisp, because the definition there of
random-list was not tail-recursive.  I wanted to fix that by replacing
that definition with the definition in books/system/random.lisp, which
uses tail recursion (and also has some lemmas proved about it).  In
order to make that change without violating an assertion in
seed-random.lisp, I reversed the accumulator before returning.
Perhaps that is unnecessary and instead the assertion(s) should be
changed.  Anyone is welcome to investigate that possibility as far as
I'm concerned (but I think it could turn painful).

Passed "make regression" in LispWorks 7.1 and "make
regression-everything" in CCL.