Fix segment address translation in X86ISA.
In 64-bit mode when the segment is not FS or GS, segment address translation
should be a no-op, i.e. it should return the effective address unmodified as a
linear address. It should not coerce the address to be canonical; the address is
checked to be canonical elsewhere, before accessing memory (through paging).

Adapt a few theorems to this change.