Improved SBCL_HOME; fixed a bug in the (raw Lisp) definition of macro ec-call1-raw.  Fixed the guard for saved-output-token-p.
Quoting :doc note-7-5:

  (SBCL only) The setting of environment variable SBCL_HOME has been
  tweaked to be more robust.  In particular, we expect it to
  unnecessary, and even inadvisable, to set SBCL_HOME manually.
  Thanks to Keshav Kini for contributing this change.

The ec-call1-raw bug was an attempt to intern into the "COMMON-LISP"
package, which isn't tolerated by all host Lisps.  In particular, it
caused errors during the build using SBCL 1.4.2.

Thanks to Mihir Mehta for reporting an error involving
saved-output-token-p when attempting to use the proof-builder.  I
found the same problem earlier in the day when attempting to call
show-accumulated-persistence, but I'd (sigh) forgotten about it!
Note: I did an :acl2-devel build and ran a successful
make devel-check.

NOTE: The ACL2 build now succeeds without errors using SBCL 1.4.2.
But it can take much longer than before: about 7 minutes on my Mac
now, but only about 1 minute 35 seconds using SBCL 1.3.2.  I'm baffled
so far as to why.  The good news is that an "everything" regression
with -j 8, based on SBCL 1.4.2, took 3 hours 34 minutes on my Mac.
That is encouraging because a similar run 23 days ago based on SBCL
1.3.2 took about 20 minutes longer -- that difference could be noise
(timing on my Mac is kind of unreliable), but at least the 1.4.2 run
wasn't slower.

Below is further explanation, from Keshav (included with permission),
of his patch.  I decided not to include this in the sources because I
think his code is very well-commented as it is; but I want to preserve
this extra explanation.

.......... start note from Keshav ..........

Hi,

Please find attached a patch.lisp file which improves ACL2's mechanism
for detecting and setting the appropriate value for the SBCL_HOME
environment variable when building ACL2 on SBCL.

The form defining *sbcl-contrib-dir* is replaced by the definition of
a new variable named *sbcl-home-dir* which has the same purpose but is
computed differently.  The body of the function save-acl2-in-sbcl-aux
is modified slightly to accommodate this.

Here are some ways in which the patch improves upon the current
situation.

- Previously, ACL2 only understood SBCL as it exists in an extracted
  binary tarball of SBCL, or an extracted source tarball of SBCL in
  which only SBCL's "make.sh" has been run.  After this patch, ACL2
  also understands SBCL as it exists in an install target directory,
  i.e. after SBCL's "install.sh" has been run as well.

  This is important because running SBCL from an install target
  directory is the standard way to run SBCL, and is how SBCL expects
  to be run -- apparently, in normal usage, manually setting SBCL_HOME
  is typically unnecessary because SBCL is able to detect the correct
  locations of things when being run from an install target directory.

  Also, in most Linux distributions, when SBCL is installed from the
  system package manager, it is placed on the user's system in the
  install target directory configuration, not the in-place
  configuration.  So this patch will help users who wish to build ACL2
  using an SBCL that was installed from their system package manager.

- Previously, if SBCL_HOME was set in the environment, ACL2 would use
  that value of SBCL_HOME without question.  After this patch, ACL2
  will ignore the value of SBCL_HOME in certain situations, namely
  when we believe that SBCL is running from an SBCL source tree ("in
  place") rather than in an install target directory *and* SBCL_HOME
  points to the directory containing sbcl.core.

  This is because such a situation is most likely to arise when an old
  version of SBCL has itself attempted to autodetect and set SBCL_HOME
  based on the location of its own core file, which makes sense when
  SBCL is being run from an install target directory, but will not
  work when SBCL is run "in place" in an SBCL source tree.

- Previously, if SBCL_HOME was *not* set in the environment, ACL2
  would refuse to build, giving the user some suggestions as to
  possible values of SBCL_HOME to set.  After this patch, ACL2 is able
  to directly use the suggestions it came up with and keep going with
  the build, without bothering the user about it.  This is an attempt
  to follow the general software engineering principle of "if you can
  do it yourself, don't tell the user to do it".

- I renamed *sbcl-contrib-dir* to *sbcl-home-dir* since that's a more
  accurate name.

- A comment about the difference between the old and new locations of
  the contrib modules in relation to the location of the sbcl.core
  file had been left behind inside the definition of
  save-acl2-in-sbcl-aux, so I moved that comment back to where it
  belongs, inside the definition of *sbcl-home-dir*.  I also added a
  note saying that the comment only applies to the "in place" case,
  and not the case when SBCL is properly installed into an install
  target directory.

- The string value of SBCL_HOME is now canonicalized as a pathname and
  checked for existence on the filesystem before being used.

- A warning is now printed if ACL2 notices that the supplied value of
  SBCL_HOME is "unexpected", in that it doesn't match the expected
  value of SBCL_HOME based on what we know about the location of
  sbcl.core.

- The selected value of SBCL_HOME is written into the saved_acl2 file
  using the ~a format specifier rather than the ~s format specifier,
  which avoids some odd issues; ~s attempts to produce a lisp form
  which can be read back into the object being formatted, which can
  theoretically lead to something other than the direct contents of
  the string being written into the file.  For example, once I saw a
  saved_acl2 end up containing something that looked like

    SBCL_HOME=#A((6) BASE-CHAR "foobar")

  rather than just

    SBCL_HOME="foobar"

  .

Ideally it would be nice if ACL2 could detect when setting SBCL_HOME
is unnecessary -- which, on modern SBCLs which are properly installed,
should be the most common scenario -- and in such a case, not include
a line setting SBCL_HOME in saved_acl2 at all.  (Note that in the
INSTALL file in the SBCL source distribution, it says that the "usual,
recommended approach" is to let SBCL find its files without setting
SBCL_HOME.)

However, I can't see a way to detect whether it's safe not to set
SBCL_HOME without actually attempting to require one of the SBCL
contrib modules during the detection process, which is presumably not
something we want to be doing in acl2-init.lisp.

Thanks,
    Keshav

.......... end note from Keshav ..........