Extend X86ISA with more 32-bit support.
Now also the TWO-BYTE-OPCODE-DECODE-AND-EXECUTE function uses INCREMENT-*IP
instead of directly incrementing the instruction pointer and checking it to be
canonical.

I had to add a somewhat inelegant local lemma to prove guards and the return
type theorem; see comment before that lemma for an explanation. We may be able
to avoid it in the future via more general and powerful rules.