Allow serialize-character-system and serialize-character to persist after make-event expansion.  Fixed books/acl2s/cgen/type.lisp for ACL2(r) by tweaking the definition of decode-ts.  Turned off waterfall-parallelism (for ACL2(p)) in books/centaur/aignet/cuts4.lisp.  Avoid stack overflows in Allegro CL for books/centaur/aignet/rwlib.lisp.
Quoting :doc note-8-0:

  The [state] global variables, serialize-character-system and
  serialize-character, are now preserved after [make-event]
  expansion.  This change for serialize-character-system allows the
  use of make-event to avoid a stack overflow when writing out a
  [certificate] file; see [set-serialize-character-system].

Pertaining to the change above (indeed, the following was the
motivation for it): books/centaur/aignet/rwlib.lisp now contains the
following form, which is mentioned in the :doc for
set-serialize-character-system as one that turns off the serialize
writer for the certificate file.

    (make-event
     (pprogn (set-serialize-character-system nil state)
             (value '(value-triple nil))))

Without that form, an attempt to certify that book, using ACL2(r)
built on Allegro CL, caused a stack overflow from ACL2 system function
ser-encode-conses.