[x86isa] Deal with syscall simulation appropriately on unsupported platforms (e.g., Windows)
Previously, on platforms other than Linux, FreeBSD, and Darwin (I'll
use Windows as the representative example here), the book
machine/syscalls.lisp would lead to a certification error because a
macro called OS (defined in that book) didn't have the appropriate
read-time conditional --- that is, something like #+(and (not linux)
(not darwin) (not freebsd))).  This is fixed now.

To clarify, the following should be possible now:

- The x86isa books should successfully certify on Windows machines.

- If the x86isa books are certified on a Windows machine, the
  following should be possible:

   1. One can reason about and execute programs compiled on Windows
      machines if they do *not* use system calls --- or more
      accurately, "pure" x86 programs not containing any OS-specific
      behavior are supported.

   2. One can reason about programs compiled on Linux, Darwin, or
      FreeBSD machines, even if they use system calls.  Use the
      os-info field of the x86 state to mimic these supported
      platforms.

   3. One can execute programs compiled on Linux, Darwin, or FreeBSD
      machines if they are "pure" x86 programs that do *not* use
      system calls.

Of course, all of this is applicable only to programmer-level mode ---
we don't do any syscall simulation in the system-level mode.

Thanks to Matt Kaufmann for bringing this issue to my attention!