Add instruction pointer operation to X86ISA.
The WRITE-*IP function "complements" the READ-*IP and INCREMENT-*IP
functions. It is not currently used by the model but it will be used by the
semantic functions of instructions to update the instruction pointer in a way
that is correct not only in 64-bit mode, but also in 32-bit mode.