Add some term utilities for substituting functions.
The new utilities FSUBLIS-FN-REC, FSUBLIS-FN, and FSUBLIS-FN-SIMPLE are variants
of the built-in SUBLIS-FN-REC, SUBLIS-FN, and SUBLIS-FN-SIMPLE: while the latter
perform some simplification, the former do not. Their relationship, and naming
convention, is analogous to FCONS-TERM and CONS-TERM (both built-in),
FSUBCOR-VAR and SUBCOR-VAR (both built-in), FSUBLIS-VAR and SUBLIS-VAR (the
former in kestrel/utilities/terms.lisp, the latter built-in), and so on.

The term utilities in kestrel/utilities/terms.lisp aim at complementing the
built-in term utilities in a well-integrated way, which motivates some choices
for naming, signatures, etc.