Extend SIB byte detection to 32-bit mode.
The function X86-DECODE-SIB-P, which decides whether a SIB is expected when
decoding an instruction, has been extended to 32-bit mode. Since the
aforementioned decision depends on whether the address size is 16 bits or not, a
boolean argument has been added to this function, saying whether the address
size is 16 bits or not.

The callers of X86-DECODE-SIB-P pass NIL as this new boolean argument for now
(thus the overall functionality of the x86 model is unchanged). The next step is
to calculate and pass the right value for this boolean argument.