Factor alignment check in X86ISA.
This is now called by the functions that read/write operands from/to memory,
when alignment checking is enabled.

It remains to decide whether this should be called by the memory read/write
functions instead, as we will have to change the operand read/write functions to
traffic in effective addresses instead of linear addresses, in order to extend
the latter functions to 32-bit mode.