Add tentative 32-bit mode extensions of some functions.
These are tentative extensions of X86-OPERAND-FROM-MODR/M-AND-SIB-BYTES,
X86-OPERAND-TO-REG/MEM, and X86-OPERAND-TO-XMM/MEM to 32-bit mode. They are
named like the existing ones but with a $ at the end.

They are "tentative" in the sense that they are currently not called by any
other part of the model, and we may change them as needed before starting to
"redirect" the callers of the old functions to these new functions.

As explained in comments in the new functions, the alignment check is a bit
awkward right now, due to the fact that the new functions operate on effective
addresses instead of linear addresses. We need to think about whether we want to
move alignment checks into the top-level memory access functions (RME-SIZE and
companions), which would have to be extended with a flag indicating whether the
memory access should be checked for alignment or not.