Fully extend MOV (opcodes 88h and 89h) to 32-bit mode.
This is now calling the extended version X86-OPERAND-TO-REG/MEM$ OF
X86-OPERAND-TO-REG/MEM, which operates on effective addresses instead of linear
addresses. The code that was possibly adding FS.base or GS.base to the address
in the semantic function has been commented out (and should be eventually
removed) because this is done by X86-OPERAND-TO-REG/MEM$ (via WME-SIZE). The
semantic function determines (the index of) the segment register to pass to
X86-OPERAND-TO-REG/MEM$ based on default (DS) and segment override prefix.

It remains to be decided whether the alignment checks should be moved to the
top-level memory access functions RME-SIZE and companions, to avoid the double
effective-to-linear translation that is being performed now. (However, this is a
very small penalty in 64-bit mode, where that translation is essentially a no-op
most of the time.)