Fixed some hard-error calls.  Put character-alistp in guard-verified :logic mode.  Supplied guards for fmt and similar functions.
Details:

Fixed some bad calls of hard-error, thanks to a bug report from Eric
Smith.

Character-alistp is now in :logic mode, guard-verified, with a
suitable forward-chaining rule,
character-alistp-forward-to-eqlable-alistp.

Quoting :doc note-8-1:

  Functions from the [fmt] family, including for example [fms] and
  [fmt-to-string], now have (incomplete) guards that, in particular,
  imply that the alist argument must satisfy character-alistp, i.e.,
  be an association list whose keys are all characters.  Thanks to
  Eric Smith for pointing out that an expression like (fmt-to-string
  "~x0" 3) could cause a raw Lisp error.

Note: I was a bit concerned that the added guards might somehow slow
down the regression.  But an "everything" regression with CCL didn't
slow down with the change.  (In fact it sped up very slightly, but I
attribute that to noise.)

; before
98713.080u 1306.812s 4:02:59.10 686.0%	0+0k 3212472+7197368io 1315pf+0w

; after
98166.056u 1248.268s 4:02:46.72 682.4%	0+0k 3430664+7201576io 1152pf+0w