Expand documentation slightly.
Say that the function to write the instruction pointer treats the 16-bit case
consistently with the writing of 16-bit values to general-purpose
registers. Even though the instruction pointer is not one of the 8
general-purpose register, given that, as already described in the documentation,
the Intel and AMD manuals are ambiguous about writing to IP, every piece of
evidence or related evidence helps and should be explicated.