Remove ASSERT! macro from l3-to-acl2 translator
@nadezhin discovered that the definition of ASSERT! in the l3-to-acl2
translator conflicted with an identically named macro in the
misc/assert book (which is also documented at :DOC assert! in the
ACL2+Books manual).

This commit solves that problem by doing away with the ASSERT! macro
in the l3-to-acl2 translator.  It was only being used in the
pattern (ASSERT! NIL <something>), which I was able to replace
everywhere with the pretty much identical
expression (PROG2$ (IMPOSSIBLE) <something>).