Manually reinstate obsolete path in generated file
Testing of the l3-to-acl2 translator is done by having a "gold"
generated output file, i.e. what one expects the translator to
produce, and then, at build time, trying to generate a new copy of the
file and compare it to the "gold" version.

However, the l3-to-acl2 translator embeds the absolute path of the l3
book into its generated output, and this can of course be different on
different systems.  Therefore, the testing scripts in the l3-to-acl2
Makefile run `diff` with an option telling it to ignore differences on
lines starting with "(INCLUDE-BOOK".

Unfortunately, if the absolute path of the l3 book is sufficiently
long, ACL2's pretty printer will insert a line break into the
INCLUDE-BOOK form, and so the trick of telling `diff` to ignore lines
starting with "(INCLUDE-BOOK" will no longer be sufficient to let the
tests pass on a system with different paths than the one on which the
"gold" version was created.

I inadvertently created exactly such a situation two commits ago, and
now I've fixed it in this commit by manually editing the generated
file mips.lisp to have the same INCLUDE-BOOK form as the Thacker
example's "gold" translator output files do.  This path doesn't seem
to be "correct", in that the directory structure of where in the ACL2
community books the l3 book is has apparently changed since whenever
the Thacker example "gold" files were generated.  But it doesn't
really matter what the INCLUDE-BOOK form is unless you want to
actually certify/load/run these files, in which case you'd probably
want to just regenerate them on your system anyway.  So I think it's
fine to have these obsolete paths in the "gold" files.  (By the way,
if the MIPS example were working, I imagine this mips.lisp file would
be put in a "gold" subdirectory, i.e. it is intended to also be a
"gold" file.)