Move NOTE defpointers to end of acl2-doc.lisp
In commit 594d28362c2ba392859e80aa00865c5dad49c462 I renamed the ACL2
1.x release note doc topics so that they use the usual naming
convention for release note doc topics.  For backwards compatibility,
I created defpointers from the old names of those topics to the new.
Now I realize that there apparently was a convention that all
defpointers in the file are in a big alphabetized list at the end, so
this commit moves those defpointers I created to where they should be.