Add a tool to save a filtered set of xdoc topics from local events. Add doc/top-fast, a new version of doc/top that uses the above tool to speed up the manual build.
Note for library maintainers: If this commit creates a "top-doc.lisp"
file in your library directory, it is for inclusion in doc/top-fast, but
you should feel free to take ownership of it and change what
documentation it collects, if you like.

The make targets for building the manual have not yet been changed.