Optional extension to number/digit conversion library.
In the library for converting between natural numbers and their big-endian and
little-endian representations in arbitrary bases, when the base is a power of 2,
the digits are unsigned bytes (in the ACL2 sense) whose size is the logarithm of
the base (e.g. if the base is 16, digits are unsigned nibbles). Thus, additional
theorems involving unsigned bytes can be proved. This optional extension of the
library supports that. It also includes a macro to generate theorems involving
unsigned byte fixtypes of specified sizes.