Improved earlier Version 8.1 fix for printing more information when a backchain-limit it exceeded.  Eliminated duplicate definition of symbol-alistp
Specifically regarding the printing fix: the message wasn't good when
the failure was due to a backchain-limit on the hypothesis of a rule.
Thanks to Eric Smith for sending an example that illustrates the
problem.  Here is a simpler example.

(defstub f1 (x) t)
(defstub f2 (x) t)
(defaxiom f1->f2
  (implies (f1 x)
           (f2 x))
  :rule-classes ((:rewrite :backchain-limit-lst (0))))
(brr t)
(monitor 'f1->f2 t)
(thm (f2 x))