Fixed build for SBCL 1.4.7.  Extended a comment with an example showing the recent type-prescription improvement.
Quoting :doc note-8-1:

  (SBCL only) ACL2 has been updated so that it builds on recent SBCL
  versions.  In particular, the build was broken for SBCL 1.4.7, as
  SBCL changed the ``RDTSC'' timing capability used in [memoization].
  Thanks to Keshav Kini for help with this issue, which has been
  resolved in ACL2 source file memoize-raw.lisp, as explained in the
  comment there about ``read-cycle-counter''.

Also, included an example simplified by J from an example I sent him,
in a comment in the source function definition of
type-set-and-returned-formals.

In SBCL, we now avoid loading "fastnumio" Quicklisp support (sorry,
I'm not even sure how to say that!)  because there are errors that I
couldn't see how to avoid with the current version of SBCL, 1.4.7.
See a comment in books/quicklisp/fastnumio-raw.lsp for various
attempts I made to overcome this problem.  Perhaps someone familiar
with Quicklisp can overcome this problem; otherwise, a note should
probably be added to :doc note-8-1-books, in books/doc/relnotes.lisp.
Note that this didn't interfere with building the manual.