Fix crash in EXPLAIN-FN
Previously, EXPLAIN-FN assumed that the only way that a rule could
have failed to apply was that one of its hypotheses failed to be
relieved.  Perhaps that's because EXPLAIN-FN was originally only meant
to deal with rewrite rules, before I changed WHY and WHY-EXPLAIN to
handle other rule types in #754.  (However, even rewrite rules can
sometimes fail for other reasons, such as because of backchain
limits.)

This commit makes EXPLAIN-FN a bit more robust; now it checks if the
failure reason is of the correct form before trying to extract a
term (the term the unrelieved hypothesis rewrote to) from it.  This
also means it's now unnecessary to run :OK-IF (BRR@ :WONP) before
calling EXPLAIN-FN.