Rename some event-level evaluation utilities.
This was agreed upon in an email discussion with Matt Kaufmann and David Rager.

The following renaming have been performed:

ENSURE-ERROR      --> MUST-FAIL-WITH-ERROR
ENSURE-SOFT-ERROR --> MUST-FAIL-WITH-SOFT-ERROR
ENSURE-HARD-ERROR --> MUST-FAIL-WITH-HARD-ERROR
THM?              --> MUST-PROVE
NOT-THM?          --> MUST-NOT-PROVE

This way, all the utilities in misc/eval.lisp start with MUST-, and the fact
that the old ENSURE-(SOFT/HARD-)ERROR utilities are specialized versions of
MUST-FAIL is more explicit.

The old names have been kept as deprecated synonyms that will be available in
the next release. They will be removed in the following release (or perhaps
after that).