Don't use require_relative gem on Ruby >= 1.9
In the README for the require_relative gem, it's recommended to
condition the requiring of the gem on the Ruby version being 1.8.x,
because Ruby 1.9.x and higher (i.e. all Ruby versions released in the
last five years) already natively support the require_relative
command.  This commit means that people using Ruby versions 1.9 and
higher won't need to go out of their way to install the
require_relative gem in order to build the VL kit.

Another option would be to just remove these "require
'require_relative'" lines and drop support for Ruby 1.8.