[x86isa] Fleshed out vex-decode-and-execute a bit
Though support for VEX-encoded instructions is far from complete right
now, the function vex-decode-and-execute checks some conditions during
instruction decoding that will lead to a #UD exception.

Also, the vex-prefixes structure is now added as an input to the
functions two-byte-opcode-decode-and-execute and
three-byte-opcode-decode-and-execute.  The value of vex-prefixes is
zero if these two opcode maps have been reached via the one-byte
opcode map (i.e., when no VEX prefixes are present). Otherwise, it
will be the value computed by the function vex-decode-and-execute.