Add theorems about string=>nats and chars=>nats
A comment that applies to both of these rules: a stronger version with
equal in the place of iff happens to be true, and that's what I was
using in my books before I moved this here - but introducing that rule
here would require either the use of heavy arithmetic libraries to
reason about <, or else a rule that relates iff and equal in a boolean
context. I am loath to do either of these things in this book, hence
the weaker version.