Extend and improve DEFBYTE.
DEFBYTE now accepts a keyword argument :PARENTS to set the XDOC parents of the
generated fixtype of bytes. All the other generated XDOC topics are now under
this topic, directly or indirectly: the recognizer and fixer are directly under
the fixtype, and the fixtype for lists is directly under the fixtype for bytes.
FTY::DEFLIST then puts its recognizer, fixer, and equivalence directly under the
fixtype for lists.

Also make minor changes to the XDOC of the macro itself.

Also create a new topic DEFBYTE-INSTANCES, and put all the predefined DEFBYTE
instances under it, using the newly introduced :PARENTS argument.