Extend DEFBYTE and change a default.
Add a keyword argument :PRED to customize the generated recognizer for the
bytes. The name ':PRED' is consistent with other FTY and STD macros. If this
argument is NIL (which is the default, i.e. if this argument is not supplied),
the predicate is obtained by adding -P after the type name. This default is also
consistent with other FTY and STD macros, but note that if :TYPE is also the
default (i.e. UBYTE<size> or SBYTE<size>), then the CLtL2 convention is to add
just P (without the dash). So the calls in existing files have been changed to
have an explicit :PRED argument.