Extend DEFBYTE.
Now the SIZE argument may be not just an explicit positive integer, but also a
named constant whose value is a positive integer, or a call of a nullary
function that provably denotes a positive integer.

In the latter case, extra local theorems are generated in the ENCAPSULATE to
(attempt to) ensure that the fixtype proofs will go through.