Add a SET-SIZE operation on lists.
This returns the size of a set represented as a list, i.e. the number of the
list's unique elements. It includes variants corresponding to the equality
variants (EQ, EQL, and EQUAL), in the same way as several other list operations,
e.g. MEMBER, SUBSETP, etc. It also comes with several theorems.

As pointed out in a discussion on this topic in the ACL2 Help mailing list, an
operation like this has been already defined in some other books, but it is part
of a separate library of sets (represented as lists). I defined this new
operation so that it can be readily integrated with the built-in and STD
operations on lists (and sets represented as lists); in particular, as mentioned
above, this operation supports the equality variants.

This operation is defined as LEN of REMOVE-DUPLICATES. As also pointed out in
the aforementioned discussion in the ACL2 Help mailing list, one could directly
use (LEN (REMOVE-DUPLICATES ...)) instead of introducing SET-SIZE. But having
tried that earlier, it looked somewhat verbose to me (especially when -EQUAL is
added to REMOVE-DUPLICATES): a single, shorter function could be useful to
improve readability, at least for some users.