Extend and improve the ABNF library a bit.
Add an argument to the LANGUAGEP predicate to specifies the top-level symbol(s)
of interest -- ABNF grammars have no explicit notion of axiom (or start symbol).
Given that, this formulation is more consistent with textbook-style formal
grammars.

Introduce a new predicate TERMINAL-STRING-FOR-RULES-P that is defined in the
same way as the previous definition of LANGUAGEP, i.e. to recognize all the
terminal strings of any rule names. This is still useful to express grammar
properties about all the terminal strings consisting of natural numbers in
certain sets (e.g. ASCII codes). So now the ABNF grammar operations related to
this are expressed in terms of this newly introduced predicate instead of the
older LANGUAGEP. Also the wording in some of the documentation has been changed
from talking about `language' to talking about `terminal strings'.