[x86isa] Created a separate dispatch for VEX-encoded instructions
The function vex-decode-and-execute now accounts for all the allowed
configurations of the VEX prefixes (as listed for each VEX-encodable
instruction in the description pages of the Intel Manual, Vol. 2,
Chapters 3, 4, 5).  Any instruction encoding not covered by this
function throws a #UD.

Also, removed the :next-byte field from vex-prefixes and evex-prefixes
structures.