[x86isa] Account for ModR/M's reg for VEX dispatch
The decoding and identification of instructions BLSR, BLSMSK, and BLSI
relies on the reg field of the ModR/M byte --- the VEX opcode maps and
dispatch functions acount for that now.  Also, since all VEX-encoded
instructions require a mandatory modr/m prefix, the
vex-decode-and-execute function now fetches the modr/m byte (and sib
byte, if necessary).