[x86isa] Fixed a bug in the ModR/M detection for INS/OUTS instructions
This was due to a bad reading of the Intel manual's opcode maps: "DX"
in the addressing info stands for the DX register, not "D x" --- the
former does not imply the presence of a ModR/M byte, while the latter
does.

Thanks to Rob Sumners for finding this bug.