Sped up a book whose certification had become slow after the recent change to keep LET expressions on right-hand sides of rewrite rules.
For the old version of the book, certification time with the new ACL2
took more than 18x the time using the old ACL2.  For the new version
of the book, the times are comparable, and a bit faster than the old
version of the book even using the old ACL2 for both.

Details:

; new ACL2, old book
2126.020u 4.928s 35:30.47 100.0%	0+0k 3208+10656io 0pf+0w
dunnottar:/projects/acl2/temp/books/centaur/vl/mlib%

; new ACL2, new book
99.624u 1.024s 1:41.82 98.8%	0+0k 3208+7656io 0pf+0w
dunnottar:/projects/acl2/temp/books/centaur/vl/mlib%

; old ACL2, old book
116.088u 1.120s 1:58.26 99.1%	0+0k 3024+7928io 0pf+0w

; old ACL2, new book
96.180u 1.088s 1:38.42 98.8%	0+0k 3208+7664io 0pf+0w