Make CHARACTER-UTILITIES a subtopic of STRING-UTILITIES.
This is the first step towards combining these two collections of utilities,
since characters and strings are closely related. This is consistent with the
STD/STRINGS library, which includes utilities not only for strings proper, but
also for characters.

These combined string/character utilities will be refactored into smaller
files/components soon.