Improve Bitcoin's Base58Check.
Require the version to be non-empty (fixing it to (0) if empty).

Prove that BASE58CHECK-ENCODE fixes its arguments to their types, including
fixing an empty version.

Add a congruence theorem about SHA-256, needed for the above proofs.

Add some library congruence theorems, needed for the above proofs. (This will be
moved to more appropriate libraries.)