Add predicate for valid code segment descriptors in 32-bit mode.
This is very similar to the one for 64-bit mode, but requires CS.L = 0 and puts
no constraints on CS.D.

Also add an XDOC topic for 32-bit mode segmentation, "parallel" to the existing
one for 64-bit mode.

Perhaps these could be merged, e.g. by having an extra PROC-MODE argument for
the recognizers.