Fix bug in JMP near.
The calculation of the updated hidden part of the CS segment should be just the
"else" branch of the previously existing conditional: the test for GDT vs. LDT
is not relevant at this point in the code, since it has already been taken care
of.

There were two instances of this issue.