Continue extending JMP far towards 32-bit mode.
Add check that, in 32-bit mode, the new offset is within the new code segment's
limit.

At this point JMP far's treatment of code segments should be fully extended to
32-bit mode. It remains to extend the treatment of call gates, and to add a
treatment of task gates and task state segments -- the latter two are not
allowed in 64-bit mode and so they were not present in the mode of JMP far.