Add AIJ and ATJ.
AIJ (= ACL2 In Java) is a deep embedding in Java of an executable,
side-effect-free, non-stobj-accessing subset of the ACL2 language without
guards.

ATJ (= ACL2 To Java) is a simple Java code generator that turns ACL2 functions
into AIJ representations that are evaluated by the AIJ interpreter.

AIJ and ATJ are described in a paper for the upcoming ACL2 Workshop. They also
have their own extensive XDOCumentation (and Javadoc for the Java code).

Only the ACL2 files are part of the overall build for now. The Java files can be
compiled and run via shell scripts. These will likely be replaced by makefiles
at some point.