[x86isa] Properly detect mandatory prefixes of opcodes with extensions
For all non-AVX opcodes, we used to compute the mandatory prefix first
and then based on the mandatory prefix, we would determine if a modr/m
byte was required.  This order works for one- and three-byte (both
0F-38 and 0F-3A) opcodes, but not for two-byte opcodes.

Specifically, for two-byte simple opcodes without extensions (i.e.,
those NOT in *opcode-extensions-by-group-number*) and for "compound"
opcodes, the above order works, but for simple opcodes with
extensions, determining the mandatory prefix can hinge on the modr/m
byte's fields.  In that case, we need to fetch the modr/m byte before
we can determine the right mandatory prefix.

This commit makes fixes to this effect.  Only the dispatch of (legacy)
two-byte opcodes is affected.