Updated my workshop contribution, including creation of a static copy, books/workshops/2018/kaufmann/.
One of the reviewers of the corresponding ACL2 Workshop 2018 paper
suggested "an appendix that is attached to some archive-hosted
preprint".  Creation of books/workshops/2018/kaufmann/ is less than
that, but at least it preserves a copy of the supporting materials as
of the time of the workshop (or more precisely, 10/1/2018).