Tweaked the last part of :doc developers-guide-contributing.  Improved directed-untranslate handling of let*.
Alessandro Coglio pointed out that, given the explanatory text, the
final "git checkout master" belongs at the beginning of the final code
block, not the end of the code block immediately preceding that one.
So I made that change.  (Of course, the entire final sequence of four
commands is the same either way; but the new version is clearer, and
in accordance with the text.)