Temporarily add slides for what's-new-books talk.
This is for the "What's New in the Community Books" talk at the upcoming ACL2
Workshop. As discussed with other contributors, since it involves contributions
from several people, the git repo is a convenient place for this (small) file,
which we can remove after the Workshop.

To the contributors that I previously sent email to: Please add/fix/edit as
needed. The format should be straightforward. Please have a look at the (few)
macros towards the top of the file, and use them. Note that if you want to
include a code block you'll have to add [fragile] just after \begin{frame}
(there's an instance of this in the file). Please make sure that pdflatex
processes the file correctly before pushing changes. Thanks.