Added support to WET for printing set-iprint suggestion.  Tweaked ccl installation instructions.
Thanks to Mihir Mehta for suggesting this enhancement to WET.

Thanks to Mertcan Temel for the tweak to the ccl installation
instructions.
NOTE: I expect to tweak them further later today (not sure), as
Mertcan and I continue to try to get to the bottom of the issue with
CCL and MacOS 10.14 (Mojave).