Improved modeling of stack pointer increment/decrement.
This operation is now modular -- 64, 32, or 16 bits, based on the current mode
aand on the SS.B bit (i.e. the "big" bit of the stack segment). Even though the
Intel and AMD manuals do not explicitly say that this operation is modular, they
do not say so for other operations either (ADD), since it should be expected.

Various (non-official) resources found on the web mention the fact that, with
expand-down stack segment, the stack pointer may be initialized to 0 (which is
not a valid effective address in an expand-down segment) and that the first push
will (modularly) decrement it and produce a valid effective address. This
further validates the fact that stack pointer operations should be modular.

The change affects the function ADD-TO-*SP. I had to modify two of the theorems
that accompany that function. I also had to adapt one proof of 64-bit programs,
namely the one for zero copy in the marking view: I added a local theorem saying
that, under the condition that a stack pointer is canonical (which is part of
the hypotheses under which the program is verified), removes the I64 conversion
now produced by ADD-TO-*SP (this conversion makes the operation modular; recall
that the model uses signed addresses in 64-bit mode).